Seamless changing between Service List, Map, Filtering views; Remove limit for visible services on the map; Add text (map, filter) to buttons with icons; Replace label "Directory" with "Services".
