+++ .eslintrc
@@ -38,6 +38,7 @@
     "no-this-before-super": 2,
     "no-multi-spaces": 1,
     "no-unused-vars": [2, {"argsIgnorePattern": "dispatch|error|response"}],
+    "object-curly-spacing": [2, "never"],
     "object-shorthand": [2, "always"],
     "prefer-spread": 1,
     "prefer-template": 1,
