+++ src/data/Services.js
@@ -49,7 +49,4 @@ export default class Services extends Component {
 
         return pagedResults;
     }
-
-
-
 }
