+++ src/data/index.js
@@ -1,3 +1,3 @@
-export { default as News } from './News';
-export { default as Services } from './Services';
-export { default as Presence } from './Presence';
+export {default as News} from './News';
+export {default as Services} from './Services';
+export {default as Presence} from './Presence';
