+++ src/reducers/index.js
@@ -1,4 +1,4 @@
-import { combineReducers } from 'redux';
+import {combineReducers} from 'redux';
 
 import region from './region';
 import language from './language';
